requirements.txt
flask==3.0.3
pandas==2.3.3
pillow==10.4.0
requests==2.32.3
gunicorn==22.0.0
python-dotenv
fpdf2==2.8.3
